Added EarlyWhiteBalance, CorrectSaturated and ClipMode Options, and other corrections
-CorrectSaturated: correct colours of saturated highlights so that they appear white after the white balance (recovers some detail in the highlights).
To test the effect, compare the result using options (OptionalTasks={'ColourCorrect'}, ColourCompatibility=false) with that of (OptionalTasks={'ColourCorrect'}, ColourCompatibility=false, CorrectSaturated=true).

-ClipMode is either 'hard' (previous behaviour), 'soft' (new soft saturation function) or 'none' (no clipping).
The default is 'soft' when CorrectSaturated is true, and 'hard' otherwise (highlights retrieved with CorrectSaturated are better preserved with soft clipping).
The mode 'none' divides the output LF by the maximul value MaxLum to prevent any clipping in the integer conversion. MaxLum is saved in metadata. To recover the correct brightness, the output must be multiplied by MaxLum (after conversion to float data).

-EarlyWhiteBalance (default=false) Performs the white balance before the call to LFColourCorrect. It applies white balance even if the 'ColourCorrect' optional task is not specified.
To test the effect, compare the result using options (OptionalTasks={'ColourCorrect'}, ColourCompatibility=false) with that of (OptionalTasks={'ColourCorrect'}, ColourCompatibility=false, EarlyWhiteBalance=true).
The difference of EarlyWhiteBalance is subtle, but it should give slightly reduced demosaicing artifacts, and better preserved colours.

-Other changes/corrections:
  -Disabled the change of saturationLevel in LFColourCorrect when the ColourCompatibility option is off. It was causing too bright images and unexpected change of brightness when switching the new EarlyWhiteBalance option on or off.
So previous tests with corrections of colours and exposure should be run again for correct exposure.
  -Always desactivate NormaliseWIExposure and NormaliseWIColours when ColourCompatibility is true (WI normalisation does not make sense in this case).
But their default is now set to true otherwise, so setting ColourCompatibility to false gives the intended colours without adjusting other options.