Added Resampling methods 'barycentric' and 'none'
-'barycentric' uses and adapts code from the KAIST decoding toolbox (Cho et al. "Modeling the calibration pipeline of the Lytro camera for high quality light-field image reconstruction",ICCV2013).
It is slow but generates larger images.
-'none' is a "non-destructive" mode from the paper (Le Pendu and Smolic,"High Resolution Light Field Recovery with Fourier Disparity Layer Completion, Demosaicing, and Super-Resolution",ICCP2020).
It extract a list of views within the disk aperture (unreliable corner views are not extracted) and without lossy interpolations.
It desactivates the colourMatrix transform to avoid mixing known data on the original RAW image with demosaiced data.
It also extracts a weight map per RGB component per view (zero weight indicate missing pixels that were not paired with an original pixel in the RAW image), and the following metadata:
  -DecodeOptions.Sampling.ST: S, T angular coordinates of each view.
  -DecodeOptions.Sampling.HexShiftStart: same as LensletGridModel.FirstPosShiftRow, required to perform hexagonal to square resampling later.
  -DecodeOptions.Sampling.SubPixAccuracy: space between 2 consecutive extracted views (measured in pixels on the RAW image).
This mode can be parameterized with 2 decoding options :
  -LensletBorderSkip: number of pixels to skip at the border of each lenslet (default=1.5 for Illum and F01).
  -NumViewsDiameter : number of extracted views on the diameter of the disk of views (default=25 for Illum, 17 for F01).

The new modes can be tested by setting DecodeOptions.ResampMethod='barycentric' or DecodeOptions.ResampMethod='none'.
Both 'barycentric' and 'none' are incompatible with the decoding mode WeightedInterp (automatically desactivated in this case).
The mode 'none' is also incompatible with eslf output (it generates 3-channel masks).