Added hot pixel correction with new option HotPixelCorrect (false by default)
For Lytro Illum, the hot pixel's positions are read directly from the HOTPIXEL.BIN file.
For Lytro F01, the hot pixel's positions are detected from the 2 black images provided in calibration files.
Currently, hot pixels detection is in LFDecodeLytroImage. But it should probably be moved to calibration, and save a .json or .bin file, to avoid recomputing each time, and to make things consistent between cameras.

To avoid changing inputs of the function LFDecodeLensletImageDirect, I added the hot pixels indices in DecodeOptions and removed them after the correction is done.